+++ examples/src/main.cpp
@@ -14,6 +14,7 @@
 #include "scenes/TriangleScene.h"
 #include "scenes/RectangleScene.h"
 #include "scenes/RectangleCameraScene.h"
+#include "scenes/CubeScene.h"
 
 // NOTE: Initialize pointer to zero so that it can be initialized in first call to getInstance
 // Because main is the first cpp file to compile and read we instantiate SceneManager here
@@ -30,7 +31,9 @@ int main(int argc, char *args[]) {
     game.getSceneManager().addScene("triangle", std::make_shared<TriangleScene>());
     game.getSceneManager().addScene("rectangle", std::make_shared<RectangleScene>());
     game.getSceneManager().addScene("rectangle_camera", std::make_shared<RectangleCameraScene>());
-    game.getSceneManager().setActiveScene("rectangle_camera");
+    game.getSceneManager().addScene("cube", std::make_shared<CubeScene>());
+
+    game.getSceneManager().setActiveScene("cube");
 
     game.run(); // run the game loop